+++ app/schedulers.py
@@ -2,6 +2,7 @@
 All functions that need to be scheduled are here.
 """
 
+import configparser
 # Build-in modules
 import logging
 import os
@@ -10,7 +11,7 @@ import os
 import requests
 
 # Local modules
-from app.shared_variables_ import currency_info
+from app.queries.query_currency import currency_info
 
 
 def get_currency():
@@ -18,7 +19,14 @@ def get_currency():
     Go fetch the currency Dollar x Brazilian Real
     """
 
-    alpha_vantage_apikey = os.environ['ALPHA_VANTAGE_KEY']
+    config = configparser.ConfigParser()
+    config.read_file(open('config.ini'))
+
+    if 'CLOUD' not in os.environ:
+        # If the application is running locally, use config.ini anf if not, use environment variables
+        alpha_vantage_apikey = config['ALPHA_VANTAGE_KEY']['key']
+    else:
+        alpha_vantage_apikey = os.environ['ALPHA_VANTAGE_KEY']
 
     # replace the "demo" apikey below with your own key from https://www.alphavantage.co/support/#api-key
     url = 'https://www.alphavantage.co/query?function=CURRENCY_EXCHANGE_RATE&from_currency=USD&to_currency=BRL&apikey' \
@@ -27,7 +35,7 @@ def get_currency():
     try:
         r = requests.get(url)
         data = r.json()
-        currency_info.dollar_brl_info = data['Realtime Currency Exchange Rate']
+        currency_info.dollar_brl_values = data['Realtime Currency Exchange Rate']
 
     except requests.exceptions as e:
         logging.exception(e, exc_info=False)
